Fixed previous sim loading, fixed EField, QSPS, added access functions to QSPS, removed attribute names as able to be specified - now Simulation specifies them, added EField API function, changed BField API funciton to take a double string instead of a single double, prepped for tracking escape and incident time.  Still need to fix Satellite sparse array functions - causes endless loops when data is not sparse.

+++ include/utils/load.h
@@ -26,7 +26,7 @@ namespace utils
 					std::vector<double> read;
 					fileIO::readDblBin(read, folder + "/" + partName + "_" + attrNames.at(attr) + ".bin");
 					data_m.push_back(read);
-					if (attrNames_m.at(attr).size() > attrsize) { attrsize = attrNames_m.at(attr).size(); }
+					if (attrNames_m.at(attr).size() > attrsize) { attrsize = (int)attrNames_m.at(attr).size(); }
 				}
 
 				for (int attr = 0; attr < attrNames.size(); attr++)
